+++ src/main/java/com/mosaicchurchaustin/oms/services/OrderService.java
@@ -75,12 +75,12 @@ public class OrderService {
         return orderHistoryRepository.findAllByOrderEntityId(pageable, orderId).getContent();
     }
 
-    public Page<OrderEntity> getOrders(final Pageable pageable, final List<String> statusFilters, final String customer) {
+    public Page<OrderEntity> getOrders(final Pageable pageable, final List<String> statusFilters, final String customer, final String customerUuid) {
         final List<OrderStatus> statusList = (statusFilters != null && !statusFilters.isEmpty())
                 ? statusFilters.stream().map(OrderStatus::from).toList()
                 : null;
 
-        final Specification<OrderEntity> spec = OrdersSpecification.withFilters(statusList, customer);
+        final Specification<OrderEntity> spec = OrdersSpecification.withFilters(statusList, customer, customerUuid);
 
         return orderRepository.findAll(spec, pageable);
     }